+++ tests/index.test.ts
@@ -2,7 +2,7 @@ import "jest";
 import * as index from "../src/index";
 
 test("should export package stuff", () => {
-    expect(index).toHaveProperty("CommandLinePrinter");
+    expect(index).toHaveProperty("StandardOutputPrinter");
 
     expect(index).toHaveProperty("ConfigurationBag");
     expect(index).toHaveProperty("ConfigurationItem");