+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Do not use `process.argv`, `process.env` and `process.stdout.columns` as default parameters values.
 - Rename `CommandLineSource` to `CommandLineArgumentsSource`.
+- Rename `CommandLinePrinter` to `StandardOutputPrinter`.
 
 ### Fixed